+++ contracts/contracts/DeployFactory.sol
@@ -10,7 +10,6 @@ import "./Verifier.sol";
 import "./TokenInit.sol";
 
 contract DeployFactory is TokenDeployInit {
-
     // Why do we deploy contracts in the constructor?
     //
     // If we want to deploy Proxy and UpgradeGatekeeper (using new) we have to deploy their contract code with this contract
@@ -28,14 +27,18 @@ contract DeployFactory is TokenDeployInit {
     // genesis state, as the very first account in tree is a fee account, and we need its address before
     // we're able to start recovering the data from the Ethereum blockchain.
     constructor(
-        Governance _govTarget, Verifier _verifierTarget, ZkSync _zkSyncTarget,
-        bytes32 _genesisRoot, address _firstValidator, address _governor,
+        Governance _govTarget,
+        Verifier _verifierTarget,
+        ZkSync _zkSyncTarget,
+        bytes32 _genesisRoot,
+        address _firstValidator,
+        address _governor,
         address _feeAccountAddress
     ) {
         require(_firstValidator != address(0));
         require(_governor != address(0));
         require(_feeAccountAddress != address(0));
-        
+
         deployProxyContracts(_govTarget, _verifierTarget, _zkSyncTarget, _genesisRoot, _firstValidator, _governor);
 
         selfdestruct(msg.sender);
@@ -43,16 +46,19 @@ contract DeployFactory is TokenDeployInit {
 
     event Addresses(address governance, address zksync, address verifier, address gatekeeper);
 
-
     function deployProxyContracts(
-        Governance _governanceTarget, Verifier _verifierTarget, ZkSync _zksyncTarget,
-        bytes32 _genesisRoot, address _validator, address _governor
+        Governance _governanceTarget,
+        Verifier _verifierTarget,
+        ZkSync _zksyncTarget,
+        bytes32 _genesisRoot,
+        address _validator,
+        address _governor
     ) internal {
-
         Proxy governance = new Proxy(address(_governanceTarget), abi.encode(this));
         // set this contract as governor
         Proxy verifier = new Proxy(address(_verifierTarget), abi.encode());
-        Proxy zkSync = new Proxy(address(_zksyncTarget), abi.encode(address(governance), address(verifier), _genesisRoot));
+        Proxy zkSync =
+            new Proxy(address(_zksyncTarget), abi.encode(address(governance), address(verifier), _genesisRoot));
 
         UpgradeGatekeeper upgradeGatekeeper = new UpgradeGatekeeper(zkSync);
 
@@ -72,9 +78,13 @@ contract DeployFactory is TokenDeployInit {
         finalizeGovernance(Governance(address(governance)), _validator, _governor);
     }
 
-    function finalizeGovernance(Governance _governance, address _validator, address _finalGovernor) internal {
+    function finalizeGovernance(
+        Governance _governance,
+        address _validator,
+        address _finalGovernor
+    ) internal {
         address[] memory tokens = getTokens();
-        for (uint i = 0; i < tokens.length; ++i) {
+        for (uint256 i = 0; i < tokens.length; ++i) {
             _governance.addToken(tokens[i]);
         }
         _governance.setValidator(_validator, true);
