+++ contracts/contracts/Operations.sol
@@ -6,10 +6,8 @@ pragma experimental ABIEncoderV2;
 
 import "./Bytes.sol";
 
-
 /// @title zkSync operations tools
 library Operations {
-
     // Circuit ops and their pubdata (chunks * bytes)
 
     /// @notice zkSync circuit operation type
@@ -59,19 +57,17 @@ library Operations {
         address owner;
     }
 
-    uint public constant PACKED_DEPOSIT_PUBDATA_BYTES =
+    uint256 public constant PACKED_DEPOSIT_PUBDATA_BYTES =
         OP_TYPE_BYTES + ACCOUNT_ID_BYTES + TOKEN_BYTES + AMOUNT_BYTES + ADDRESS_BYTES;
 
     /// Deserialize deposit pubdata
-    function readDepositPubdata(bytes memory _data) internal pure
-        returns (Deposit memory parsed)
-    {
+    function readDepositPubdata(bytes memory _data) internal pure returns (Deposit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = OP_TYPE_BYTES;
+        uint256 offset = OP_TYPE_BYTES;
         (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
-        (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset);   // tokenId
-        (offset, parsed.amount) = Bytes.readUInt128(_data, offset);   // amount
-        (offset, parsed.owner) = Bytes.readAddress(_data, offset);    // owner
+        (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
+        (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
+        (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
 
         require(offset == PACKED_DEPOSIT_PUBDATA_BYTES, "rdp10"); // reading invalid deposit pubdata size
     }
@@ -80,18 +76,18 @@ library Operations {
     function writeDepositPubdata(Deposit memory op) internal pure returns (bytes memory buf) {
         buf = abi.encodePacked(
             uint8(OpType.Deposit),
-            bytes4(0),   // accountId (ignored) (update when ACCOUNT_ID_BYTES is changed)
-            op.tokenId,  // tokenId
-            op.amount,   // amount
-            op.owner     // owner
+            bytes4(0), // accountId (ignored) (update when ACCOUNT_ID_BYTES is changed)
+            op.tokenId, // tokenId
+            op.amount, // amount
+            op.owner // owner
         );
     }
 
     /// @notice Check that deposit pubdata from request and block matches
     function depositPubdataMatch(bytes memory _lhs, bytes memory _rhs) internal pure returns (bool) {
         // We must ignore `accountId` and operation type because it is present in block pubdata but not in priority queue
-        uint skipBytes = ACCOUNT_ID_BYTES + OP_TYPE_BYTES;
-        bytes memory lhs_trimmed = Bytes.slice(_lhs, skipBytes , PACKED_DEPOSIT_PUBDATA_BYTES - skipBytes);
+        uint256 skipBytes = ACCOUNT_ID_BYTES + OP_TYPE_BYTES;
+        bytes memory lhs_trimmed = Bytes.slice(_lhs, skipBytes, PACKED_DEPOSIT_PUBDATA_BYTES - skipBytes);
         bytes memory rhs_trimmed = Bytes.slice(_rhs, skipBytes, PACKED_DEPOSIT_PUBDATA_BYTES - skipBytes);
         return keccak256(lhs_trimmed) == keccak256(rhs_trimmed);
     }
@@ -105,18 +101,16 @@ library Operations {
         uint128 amount;
     }
 
-    uint public constant PACKED_FULL_EXIT_PUBDATA_BYTES =
+    uint256 public constant PACKED_FULL_EXIT_PUBDATA_BYTES =
         OP_TYPE_BYTES + ACCOUNT_ID_BYTES + ADDRESS_BYTES + TOKEN_BYTES + AMOUNT_BYTES;
 
-    function readFullExitPubdata(bytes memory _data) internal pure
-        returns (FullExit memory parsed)
-    {
+    function readFullExitPubdata(bytes memory _data) internal pure returns (FullExit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = OP_TYPE_BYTES;
-        (offset, parsed.accountId) = Bytes.readUInt32(_data, offset);      // accountId
-        (offset, parsed.owner) = Bytes.readAddress(_data, offset);         // owner
-        (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset);        // tokenId
-        (offset, parsed.amount) = Bytes.readUInt128(_data, offset);        // amount
+        uint256 offset = OP_TYPE_BYTES;
+        (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
+        (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
+        (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
+        (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
 
         require(offset == PACKED_FULL_EXIT_PUBDATA_BYTES, "rfp10"); // reading invalid full exit pubdata size
     }
@@ -124,23 +118,23 @@ library Operations {
     function writeFullExitPubdata(FullExit memory op) internal pure returns (bytes memory buf) {
         buf = abi.encodePacked(
             uint8(OpType.FullExit),
-            op.accountId,  // accountId
-            op.owner,      // owner
-            op.tokenId,    // tokenId
-            op.amount      // amount
+            op.accountId, // accountId
+            op.owner, // owner
+            op.tokenId, // tokenId
+            op.amount // amount
         );
     }
 
     /// @notice Check that full exit pubdata from request and block matches
     function fullExitPubdataMatch(bytes memory _lhs, bytes memory _rhs) internal pure returns (bool) {
         // `amount` is ignored because it is present in block pubdata but not in priority queue
-        uint lhs = Bytes.trim(_lhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
-        uint rhs = Bytes.trim(_rhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
+        uint256 lhs = Bytes.trim(_lhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
+        uint256 rhs = Bytes.trim(_rhs, PACKED_FULL_EXIT_PUBDATA_BYTES - AMOUNT_BYTES);
         return lhs == rhs;
     }
 
     // PartialExit pubdata
-    
+
     struct PartialExit {
         //uint8 opType
         //uint32 accountId; -- present in pubdata, ignored at serialization
@@ -150,19 +144,17 @@ library Operations {
         address owner;
     }
 
-    function readPartialExitPubdata(bytes memory _data) internal pure
-        returns (PartialExit memory parsed)
-    {
+    function readPartialExitPubdata(bytes memory _data) internal pure returns (PartialExit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = OP_TYPE_BYTES + ACCOUNT_ID_BYTES;             // opType + accountId (ignored)
+        uint256 offset = OP_TYPE_BYTES + ACCOUNT_ID_BYTES; // opType + accountId (ignored)
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
-        offset += FEE_BYTES;                                        // fee (ignored)
-        (offset, parsed.owner) = Bytes.readAddress(_data, offset);  // owner
+        offset += FEE_BYTES; // fee (ignored)
+        (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
     }
 
     // ForcedExit pubdata
-    
+
     struct ForcedExit {
         //uint8 opType; -- present in pubdata, ignored at serialization
         //uint32 initiatorAccountId; -- present in pubdata, ignored at serialization
@@ -173,22 +165,18 @@ library Operations {
         address target;
     }
 
-    function readForcedExitPubdata(bytes memory _data) internal pure
-        returns (ForcedExit memory parsed)
-    {
+    function readForcedExitPubdata(bytes memory _data) internal pure returns (ForcedExit memory parsed) {
         // NOTE: there is no check that variable sizes are same as constants (i.e. TOKEN_BYTES), fix if possible.
-        uint offset = OP_TYPE_BYTES + ACCOUNT_ID_BYTES * 2;         // opType + initiatorAccountId + targetAccountId (ignored)
+        uint256 offset = OP_TYPE_BYTES + ACCOUNT_ID_BYTES * 2; // opType + initiatorAccountId + targetAccountId (ignored)
         (offset, parsed.tokenId) = Bytes.readUInt16(_data, offset); // tokenId
         (offset, parsed.amount) = Bytes.readUInt128(_data, offset); // amount
-        offset += FEE_BYTES;                                        // fee (ignored)
+        offset += FEE_BYTES; // fee (ignored)
         (offset, parsed.target) = Bytes.readAddress(_data, offset); // target
     }
+
     // ChangePubKey
 
-    enum ChangePubkeyType {
-        ECSDA,
-        Create2
-    }
+    enum ChangePubkeyType {ECSDA, Create2}
 
     struct ChangePubKey {
         // uint8 opType; -- present in pubdata, ignored at serialization
@@ -200,26 +188,30 @@ library Operations {
         //uint16 fee; -- present in pubdata, ignored at serialization
     }
 
-    function readChangePubKeyPubdata(bytes memory _data) internal pure
-        returns (ChangePubKey memory parsed)
-    {
-        uint offset = OP_TYPE_BYTES;
-        (offset, parsed.accountId) = Bytes.readUInt32(_data, offset);                // accountId
-        (offset, parsed.pubKeyHash) = Bytes.readBytes20(_data, offset);              // pubKeyHash
-        (offset, parsed.owner) = Bytes.readAddress(_data, offset);                   // owner
-        (offset, parsed.nonce) = Bytes.readUInt32(_data, offset);                    // nonce
+    function readChangePubKeyPubdata(bytes memory _data) internal pure returns (ChangePubKey memory parsed) {
+        uint256 offset = OP_TYPE_BYTES;
+        (offset, parsed.accountId) = Bytes.readUInt32(_data, offset); // accountId
+        (offset, parsed.pubKeyHash) = Bytes.readBytes20(_data, offset); // pubKeyHash
+        (offset, parsed.owner) = Bytes.readAddress(_data, offset); // owner
+        (offset, parsed.nonce) = Bytes.readUInt32(_data, offset); // nonce
     }
 
     // Withdrawal data process
 
-    function readWithdrawalData(bytes memory _data, uint _offset) internal pure
-        returns (bool _addToPendingWithdrawalsQueue, address _to, uint16 _tokenId, uint128 _amount)
+    function readWithdrawalData(bytes memory _data, uint256 _offset)
+        internal
+        pure
+        returns (
+            bool _addToPendingWithdrawalsQueue,
+            address _to,
+            uint16 _tokenId,
+            uint128 _amount
+        )
     {
-        uint offset = _offset;
+        uint256 offset = _offset;
         (offset, _addToPendingWithdrawalsQueue) = Bytes.readBool(_data, offset);
         (offset, _to) = Bytes.readAddress(_data, offset);
         (offset, _tokenId) = Bytes.readUInt16(_data, offset);
         (offset, _amount) = Bytes.readUInt128(_data, offset);
     }
-
 }
