+++ contracts/contracts/UpgradeGatekeeper.sol
@@ -18,24 +18,20 @@ contract UpgradeGatekeeper is UpgradeEvents, Ownable {
     Upgradeable[] public managedContracts;
 
     /// @notice Upgrade mode statuses
-    enum UpgradeStatus {
-        Idle,
-        NoticePeriod,
-        Preparation
-    }
+    enum UpgradeStatus {Idle, NoticePeriod, Preparation}
 
     UpgradeStatus public upgradeStatus;
 
     /// @notice Notice period finish timestamp (as seconds since unix epoch)
     /// @dev Will be equal to zero in case of not active upgrade mode
-    uint public noticePeriodFinishTimestamp;
+    uint256 public noticePeriodFinishTimestamp;
 
     /// @notice Addresses of the next versions of the contracts to be upgraded (if element of this array is equal to zero address it means that appropriate upgradeable contract wouldn't be upgraded this time)
     /// @dev Will be empty in case of not active upgrade mode
     address[] public nextTargets;
 
     /// @notice Version id of contracts
-    uint public versionId;
+    uint256 public versionId;
 
     /// @notice Contract which defines notice period duration and allows finish upgrade during preparation of it
     UpgradeableMaster public mainContract;
@@ -65,7 +61,7 @@ contract UpgradeGatekeeper is UpgradeEvents, Ownable {
         require(upgradeStatus == UpgradeStatus.Idle, "spu11"); // spu11 - unable to activate active upgrade mode
         require(newTargets.length == managedContracts.length, "spu12"); // spu12 - number of new targets must be equal to the number of managed contracts
 
-        uint noticePeriod = mainContract.getNoticePeriod();
+        uint256 noticePeriod = mainContract.getNoticePeriod();
         mainContract.upgradeNoticePeriodStarted();
         upgradeStatus = UpgradeStatus.NoticePeriod;
         noticePeriodFinishTimestamp = block.timestamp.add(noticePeriod);
@@ -123,5 +119,4 @@ contract UpgradeGatekeeper is UpgradeEvents, Ownable {
         noticePeriodFinishTimestamp = 0;
         delete nextTargets;
     }
-
 }
